Only restore backup if it exists.

+++ projects/cmake/build-ninja.sh
@@ -23,6 +23,8 @@ fi
 	ninja
 	cd ..
 	
-    cp GitVersion_backup.cpp ../../src/revlanguage/utils/GitVersion.cpp
-    rm GitVersion_backup.cpp
+    if [ -e  GitVersion_backup.cpp ] ; then
+        cp GitVersion_backup.cpp ../../src/revlanguage/utils/GitVersion.cpp
+        rm GitVersion_backup.cpp
+    fi
 fi
